+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+- observability: deprecation of logging levels `applicationError` and `serverError` in favor
+  of `serverError` and `clientError`.
 
 ## [1.52.0] - 2021-02-12
 ### Added